Delay creation of FlashMessenger
The constructor of the FlashMessenger tries to initialize a session with ```forum_session_start``` but the hook in it will never fire since the hooks haven't been loaded at that point. 

None of the code between the old and my proposed new positions tries to use ```$forum_flash```, so it doesn't make a difference from that perspective, but ```fn_forum_session_start_start``` hooks will only fire if ```$forum_flash = new FlashMessenger();``` is AFTER the hook loader.