Fix unique key in 'online'-table to accept ipv6 addresses
When accessing a punbb installation with ipv6, under certain
circumstances requests fail with db errors.
This has been caused by truncating the length of ident as
part of the unique id in table 'online' to 25 characters.
This is too short for ipv6 which can have a maximum length
of 40 characters.
This commit changes the length to 40 characters and therefore
fixes this issue.